pyproject.toml
[tool.poetry]
name = "python-screenshots"
version = "0.0.1"
description = ""
authors = ["Petr Stribny <petrstribny@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyppeteer = "^0.2.2"
pillow = "^7.2.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
